backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
celery[redis]
requests
pydantic
python-dotenv